Fix Windows compile error, conversion from int64 to int
TBR=yangguo@chromium.org

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@23780 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/interface-descriptors.cc
@@ -44,7 +44,7 @@ void CallInterfaceDescriptorData::Initialize(
 
 const char* CallInterfaceDescriptor::DebugName(Isolate* isolate) {
   CallInterfaceDescriptorData* start = isolate->call_descriptor_data(0);
-  int index = data_ - start;
+  size_t index = data_ - start;
   DCHECK(index < CallDescriptors::NUMBER_OF_DESCRIPTORS);
   CallDescriptors::Key key = static_cast<CallDescriptors::Key>(index);
   switch (key) {
